requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
lxml==6.0.1
MechanicalSoup==1.4.0
multidict==6.6.4
propcache==0.3.2
python-dotenv==1.1.1
requests==2.32.5
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
torch
torchvision
open_clip_torch
scikit-learn
pandas
numpy
Pillow
requests
beautifulsoup4
streamlit
timm
